utilisation des DTO et correction sur la fonctionnalités du filtre pour qu'il puisse gérer la pagination

+++ src/main/java/com/excilys/cdb/model/Company.java
@@ -33,7 +33,7 @@ public class Company {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
